Refactoring: 시멘틱 태그
