+++ public/index.html
@@ -12,6 +12,6 @@
   </head>
 
   <body>
-    <div id="root"></div>
+    <main id="root"></main>
   </body>
 </html>
