+++ src/components/Auth/SignUpForm.jsx
@@ -1,56 +1,15 @@
 import React, { useEffect, useState } from "react";
-import styled from "styled-components";
 import { useForm } from "react-hook-form";
-import { ButtonStyle } from "../common/Button/Button";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
-import _ from "lodash";
-const StyledForm = styled.form`
-  margin-top: 40px;
-  height: 100%;
-`;
-const StyledButton = styled(ButtonStyle)`
-  margin: 50px auto 20px;
-`;
-const StyledLabel = styled.label`
-  display: block;
-  text-align: left;
-  padding: 0 35px;
-  font-size: 16px;
-  color: #767676;
-  pointer-events: none;
-`;
-const StyledInput = styled.input`
-  display: block;
-  width: 322px;
-  box-sizing: border-box;
-  border: none;
-  box-shadow: 0 1px 0 0 #677880;
-  height: 46px;
-  border-radius: 4px 4px 0 0;
-  padding-top: 8px;
-  font-size: 14px;
-  margin: 0 auto 36px auto;
-  outline: none;
-  background: transparent;
-  &::placeholder {
-    color: #dbdbdb;
-  }
-  &:focus {
-    outline: none;
-    border-bottom: 1px solid #286140;
-  }
-`;
-const StyledInputContainer = styled.div`
-  position: relative;
-`;
-const StyledError = styled.small`
-  font-size: 12px;
-  color: red;
-  position: absolute;
-  bottom: -18px;
-  left: 35px;
-`;
+import {
+  StyledButton,
+  StyledError,
+  StyledForm,
+  StyledInput,
+  StyledInputContainerDiv,
+  StyledLabel,
+} from "./SignUpFormStyle";
 
 const SignUpForm = () => {
   const {
@@ -122,7 +81,7 @@ const SignUpForm = () => {
 
   return (
     <StyledForm onSubmit={handleSubmit(handleFormSubmit)}>
-      <StyledInputContainer>
+      <StyledInputContainerDiv>
         <StyledLabel htmlFor="email">이메일</StyledLabel>
         <StyledInput
           id="email"
@@ -140,9 +99,9 @@ const SignUpForm = () => {
         {errors.email && (
           <StyledError role="alert">{errors.email.message}</StyledError>
         )}
-      </StyledInputContainer>
+      </StyledInputContainerDiv>
 
-      <StyledInputContainer>
+      <StyledInputContainerDiv>
         <StyledLabel htmlFor="password">비밀번호</StyledLabel>
         <StyledInput
           id="password"
@@ -159,8 +118,8 @@ const SignUpForm = () => {
         {errors.password && (
           <StyledError role="alert">{errors.password.message}</StyledError>
         )}
-      </StyledInputContainer>
-      <StyledInputContainer>
+      </StyledInputContainerDiv>
+      <StyledInputContainerDiv>
         <StyledLabel htmlFor="passwordConfirm">비밀번호 확인</StyledLabel>
         <StyledInput
           id="passwordConfirm"
@@ -181,7 +140,7 @@ const SignUpForm = () => {
             {errors.passwordConfirm.message}
           </StyledError>
         )}
-      </StyledInputContainer>
+      </StyledInputContainerDiv>
 
       <StyledButton
         type="submit"
